move directory browsing code into its own js file
